refactor(metrics): move to prometheus folder
